+++ public/js/vue/initialize-clientside.js
@@ -2,6 +2,7 @@
 
 const createStore = require('./store').default;
 const renderLeftMenu = require('./left-menu').default;
+const renderThreadMessageFeed = require('./thread-message-feed').default;
 const setupDataBridge = require('./store/data-bridge').default;
 
 const store = createStore();
@@ -18,4 +19,9 @@ setupDataBridge(store);
 const leftMenuRootEl = document.querySelector('.js-left-menu-root');
 if (leftMenuRootEl) {
   renderLeftMenu(leftMenuRootEl, store);
+}
+
+const threadMessageFeedRootEl = document.querySelector('.js-thread-message-feed-root');
+if (threadMessageFeedRootEl) {
+  renderThreadMessageFeed(threadMessageFeedRootEl, store);
 }